Resolves #3920 by implementing @whophil's suggestion always using `vtkAppendFilter` when merging PolyData and then casting back.

The speedup is noticeable where it counts. Seems that the point merging algorithm grows quadraticly  with `vtkCleanPolyData`.

### 1M points example
```py
>>> import pyvista as pv
>>> sphere_a = pv.Sphere(theta_resolution=1000, phi_resolution=1000)
>>> sphere_b = sphere_a.copy()
>>> timeit sphere_a.merge(sphere_b, merge_points=True)
```

#### `main` branch
```
1.48 s ± 6.07 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
#### This branch
```
473 ms ± 3.39 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### 4M points example
```py
>>> import pyvista as pv
>>> sphere_a = pv.Sphere(theta_resolution=2000, phi_resolution=2000)
>>> sphere_b = sphere_a.copy()
>>> timeit sphere_a.merge(sphere_b, merge_points=True)
```

#### `main` branch
```
17.4 s ± 29.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

#### This branch
```
1.97 s ± 21.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```