@whophil 2023-04-04
LGTM, thanks for implementing @akaszynski!

I just want to add that this change is slightly more than a mere performance twiddle. PyVista will no longer use the `vtkAppendPolyData` filter at all, when it used to do so when merging all-polydata sets. Perhaps this has some other implications.