@MbBrainz 2023-06-02
There is a very important implication of this change which is definitely breaking. I have a piece of code that relies on the order of the points after the merge, which is preserved in Pyvista v0.38.1, but not in v0.39.0. 

I've spent a week to figure this out. Of course this is a unstable release and its part of the usage risk, however it would be great to have this ordering of points as part of the testsuite. 

This change should be explicitly mentioned in the documentation or fixed. Im not familiar with the exact workings of the underneath vtk funtionality, but i can imagine that the changing of the filter usage discussed above may have introduced this different behaviour. In that case, a solution would be to add a flag to the PolyData.merge functionality. Something like "preserve_order"=True