@beroda 2023-06-02
@banesullivan @akaszynski I am experiencing the same issue as @MbBrainz.

Switching to pyvista 0.39 in our code base at work caused some tests to crash since we rely on the points order for FEM applications.
Is there a workaround ? If not, do you mind if I open an issue to solve that ?
I could propose a PR if needed.