@beroda 2023-06-02
@banesullivan No worries, an edge-case is ... an edge-case :wink:

If it is fine for you, I think I will go for the `PolyData.append_polydata` solution, and add a warning to the `merge` documentation.