@banesullivan 2023-06-02
Looks like @whophil's foresight was on the nose here -- sounds beneficial to add back a method to use `vtkAppendPolyData`. @beroda go ahead with that approach.